requirements.txt
langchain
pinecone-client
openai
tiktoken
nest_asyncio
Flask-SQLAlchemy
lxml
beautifulsoup4
sentence-transformers